refactored create_tags to enable easier intergration of different sections
