+++ src/utils/__init__.py
@@ -163,3 +163,5 @@ def generate_cache_key(f, *args, **kwargs):
     # Construct the cache key using function name, args, and kwargs
     cache_key = f"{f.__name__}{kwargs}"
     return sanitize_filename(cache_key)
+
+
