+++ template/blog/blog.html
@@ -1,10 +1,13 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
+{% extends "layouts/page.html" %}
+{% block content %}
+    <section class="container-fluid">
+        <div class="card-content">
+            <div class="card-header">
+                <h1 class="card-title">{{ term|title }}</h1>
+            </div>
+        </div>
 
-</body>
-</html>
+
+    </section>
+{% endblock %}
+
