+++ template/layouts/page.html
@@ -57,7 +57,6 @@
 
 {% endblock %}
 <style>
-<style>
     /* Add this CSS to your stylesheet or in a <style> tag in your HTML */
 
     /* Hide the desktop buttons by default */
@@ -101,7 +100,7 @@
     }
 
 </style>
-</style>
+
 <script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-7134185943809269"
      crossorigin="anonymous"></script>
 </head>
